+++ vmm/src/config.rs
@@ -12,6 +12,7 @@ use std::str::FromStr;
 use std::sync::LazyLock;
 
 use api_types::{RestoredVfioConfig, VmMemoryZoneUpdateData};
+use api_types::MemoryRestoreMode;
 use arch::CpuProfile;
 use block::ImageType;
 use clap::ArgMatches;
@@ -2785,33 +2786,6 @@ where
         Ok(Some(vec![-1; invalid_fds.len()]))
     } else {
         Ok(None)
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
-pub enum MemoryRestoreMode {
-    /// Restore by eagerly copying the snapshot into guest RAM before resume.
-    #[default]
-    Copy,
-    /// Restore lazily by faulting snapshot pages into guest RAM on demand.
-    OnDemand,
-}
-
-#[derive(Debug, Error)]
-pub enum MemoryRestoreModeParseError {
-    #[error("Invalid value: {0}")]
-    InvalidValue(String),
-}
-
-impl FromStr for MemoryRestoreMode {
-    type Err = MemoryRestoreModeParseError;
-
-    fn from_str(s: &str) -> result::Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "copy" => Ok(Self::Copy),
-            "ondemand" => Ok(Self::OnDemand),
-            _ => Err(MemoryRestoreModeParseError::InvalidValue(s.to_owned())),
-        }
     }
 }
